merge: merge piedoom/restructure into optimizations branch
