+++ .idea/cody_history.xml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ChatHistory">
+    <accountData>
+      <list>
+        <AccountData>
+          <accountId value="VXNlcjozODgwNDA=" />
+          <defaultLlm>
+            <llm>
+              <model value="anthropic::2024-10-22::claude-3-5-sonnet-latest" />
+              <provider value="anthropic" />
+              <tags>
+                <list>
+                  <option value="free" />
+                  <option value="vision" />
+                  <option value="power" />
+                </list>
+              </tags>
+              <title value="Claude 3.5 Sonnet (New)" />
+              <usage>
+                <list>
+                  <option value="edit" />
+                  <option value="chat" />
+                </list>
+              </usage>
+            </llm>
+          </defaultLlm>
+        </AccountData>
+      </list>
+    </accountData>
+  </component>
+</project>
