+++ Cargo.toml
@@ -5,12 +5,8 @@ edition = "2021"
 license = "MIT OR Apache-2.0"
 authors = ["Yannik Bretschneider <me@yandrik.dev>"]
 description = "embedded-graphics based GUI framework made to be as lightweight as its namesake"
-readme = "../README.md"
+readme = "README.md"
 repository = "https://github.com/Yandrik/kolibri"
-
-[features]
-std-debug = []
-
 
 [lib]
 
@@ -18,9 +14,29 @@ std-debug = []
 
 [dependencies]
 embedded-graphics = "^0.8"
-embedded-iconoir = { version = "0.2.2", features = ["all-resolutions"] }
+embedded-iconoir = { version = "0.2.3", features = ["all-resolutions"] }
 foldhash = { version = "0.1.3", default-features = false }
-heapless = { version = "^0.7", features = ["serde", "ufmt-impl"] }
+heapless = { version = "^0.8", features = ["serde"] }
 
 [dev-dependencies]
-embedded-graphics-simulator = "0.5.0"
+embedded-graphics-simulator = "0.7.0"
+profont = "0.7.0"
+ibm437 = "0.3.3"
+az = "1.2"
+ansi-parser = { version = "0.9.1", default-features = false }
+object-chain = "0.1"
+
+[[example]]
+name = "basic-example"
+
+[[example]]
+name = "basic-example-incremental-redraw"
+
+[[example]]
+name = "theming"
+
+[[example]]
+name = "motion-scheduler"
+
+[[example]]
+name = "keyboard"
