+++ README.md
@@ -9,6 +9,8 @@ Kolibri is designed to be dead simple.
 > All examples used in this section can also be found in the `examples` folder.
 > Large amounts of documentation aren't available yet, but will be added before Kolibri reaches 0.1.0
 
+To get started, add Kolibri to your project using `cargo add kolibri-embedded-gui`, create an `Ui` with a mutable reference to your `embedded-graphics` `DrawTarget`, and you're ready to go!
+
 ### Super simple, hyper-effective
 
 Iterate super-quickly, design GUIs like you would write a word document, and run it on literally anything - in minutes.
