--- examples/Cargo.toml
@@ -1,34 +1,0 @@
-[package]
-name = "kolibri-gui-examples"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[[bin]]
-name = "basic-example"
-
-
-[[bin]]
-name = "basic-example-incremental-redraw"
-
-[[bin]]
-name = "theming"
-
-[[bin]]
-name = "experimenting"
-
-[[bin]]
-name = "motion-scheduler"
-
-[[bin]]
-name = "keyboard"
-
-[dependencies]
-
-kolibri-embedded-gui = { path = "../kolibri-embedded-gui", features = ["std-debug"] }
-embedded-graphics-simulator = "0.5.0"
-embedded-graphics = "0.8.0"
-profont = "0.7.0"
-ibm437 = "0.3.2"
-
