+++ examples/basic-example-incremental-redraw.rs
@@ -1,23 +1,15 @@
 use embedded_graphics::geometry::Size;
 use embedded_graphics::mono_font::ascii;
-use embedded_graphics::pixelcolor::{Rgb565, RgbColor};
-use embedded_graphics::prelude::{Point, WebColors};
-use embedded_graphics::primitives::{Circle, PrimitiveStyle, StyledDrawable};
-use embedded_graphics::text::Text;
+use embedded_graphics::pixelcolor::Rgb565;
+use embedded_graphics::prelude::Point;
 use embedded_graphics_simulator::sdl2::MouseButton;
 use embedded_graphics_simulator::{
-    BinaryColorTheme, OutputSettingsBuilder, SimulatorDisplay, SimulatorEvent, Window,
+    OutputSettingsBuilder, SimulatorDisplay, SimulatorEvent, Window,
 };
 use kolibri_embedded_gui::button::Button;
-use kolibri_embedded_gui::checkbox::Checkbox;
-use kolibri_embedded_gui::icon::IconWidget;
-use kolibri_embedded_gui::iconbutton::IconButton;
-use kolibri_embedded_gui::icons::{size12px, size24px, size32px};
 use kolibri_embedded_gui::label::Label;
-use kolibri_embedded_gui::prelude::*;
-use kolibri_embedded_gui::smartstate::{Smartstate, SmartstateProvider};
-use kolibri_embedded_gui::spacer::Spacer;
-use kolibri_embedded_gui::style::{medsize_rgb565_debug_style, medsize_rgb565_style};
+use kolibri_embedded_gui::smartstate::SmartstateProvider;
+use kolibri_embedded_gui::style::medsize_rgb565_style;
 use kolibri_embedded_gui::ui::{Interaction, Ui};
 
 fn main() -> Result<(), core::convert::Infallible> {
