+++ src/ui.rs
@@ -302,7 +302,7 @@ impl Placer {
 
     /// Check whether a size is in bounds of the widget (<= widget_size)
     fn check_bounds(&self, pos: Size) -> bool {
-        pos.width as u32 <= self.bounds.width && pos.height <= self.bounds.height
+        pos.width <= self.bounds.width && pos.height <= self.bounds.height
     }
 }
 
@@ -326,7 +326,7 @@ impl<'a, COL: PixelColor, DRAW: DrawTarget<Color = COL>> Painter<'a, COL, DRAW> 
     }
 
     fn start_drawing(&mut self, area: &Rectangle) {
-        if let Some(_) = self.framebuf {
+        if self.framebuf.is_some() {
             panic!("Framebuffer is already in use!");
         }
 
@@ -380,7 +380,7 @@ impl<'a, COL: PixelColor, DRAW: DrawTarget<Color = COL>> Painter<'a, COL, DRAW> 
         let target: &'b mut DRAW = self.target;
         let mut subpainter = Painter::new(target);
 
-        if matches!(self.framebuf, Some(_)) {
+        if self.framebuf.is_some() {
             panic!("Cannot create subpainter when framebuf is in use!");
         }
 
@@ -416,7 +416,7 @@ impl<COL: PixelColor, DRAW: DrawTarget<Color = COL, Error = ERR>, ERR> DrawTarge
 }
 
 /// Interaction with the UI
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Default)]
 pub enum Interaction {
     /// A click event (mouse, touch, etc. down)
     Click(Point),
@@ -428,13 +428,8 @@ pub enum Interaction {
     /// Generally not applicable to touch screens.
     Hover(Point),
     /// No interaction
+    #[default]
     None,
-}
-
-impl Default for Interaction {
-    fn default() -> Self {
-        Interaction::None
-    }
 }
 
 impl Interaction {
@@ -465,7 +460,7 @@ where
 }
 
 // getters for Ui things
-impl<'a, DRAW, COL> Ui<'a, DRAW, COL>
+impl<DRAW, COL> Ui<'_, DRAW, COL>
 where
     DRAW: DrawTarget<Color = COL>,
     COL: PixelColor,
@@ -652,8 +647,7 @@ where
 
     /// For now, only stub method.
     pub fn allocate_exact_size(&mut self, desired_size: Size) -> GuiResult<InternalResponse> {
-        let allocated = self.allocate_space(desired_size);
-        allocated
+        self.allocate_space(desired_size)
     }
 
     pub fn allocate_space(&mut self, desired_size: Size) -> GuiResult<InternalResponse> {
@@ -689,7 +683,7 @@ where
 }
 
 // Clearing impls
-impl<'a, COL, DRAW> Ui<'a, DRAW, COL>
+impl<COL, DRAW> Ui<'_, DRAW, COL>
 where
     DRAW: DrawTarget<Color = COL>,
     COL: PixelColor,
@@ -812,7 +806,7 @@ where
 
 // SubUI impl
 
-impl<'a, COL, DRAW> Ui<'a, DRAW, COL>
+impl<COL, DRAW> Ui<'_, DRAW, COL>
 where
     DRAW: DrawTarget<Color = COL>,
     COL: PixelColor,
@@ -845,7 +839,7 @@ where
             let mut sub_ui = Ui {
                 painter,
                 bounds,
-                style: self.style.clone(),
+                style: self.style,
                 interact: self.interact,
                 placer,
                 cleared: false,
@@ -865,7 +859,7 @@ where
             let mut sub_ui = Ui {
                 painter,
                 bounds: self.bounds,
-                style: self.style.clone(),
+                style: self.style,
                 interact: self.interact,
                 placer: self.placer.clone(),
                 cleared: false,
@@ -948,7 +942,7 @@ where
 
 // debug drawing impl
 
-impl<'a, COL, DRAW> Ui<'a, DRAW, COL>
+impl<COL, DRAW> Ui<'_, DRAW, COL>
 where
     DRAW: DrawTarget<Color = COL>,
     COL: PixelColor,
